+++ extensions/glimpse/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Glimpse Changelog
 
+## [Fix Model Downloads] - {PR_MERGE_DATE}
+
+- Switch Speech Model sends downloads to the Models page in Glimpse, instead of failing.
+
 ## [Initial Release] - 2026-07-27